requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
asttokens==2.0.5
audioread==2.1.9
autopep8==1.4.4
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
executing==0.8.3
filelock==3.6.0
fonttools==4.33.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.3
librosa==0.9.1
llvmlite==0.38.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numba==0.55.2
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pbr==5.8.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.1.1
pipx==1.0.0
platformdirs==2.5.1
pooch==1.6.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.5.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.7
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.1
pytz==2022.1
pywin32==303
requests==2.27.1
resampy==0.2.2
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.2.4
stack-data==0.3.0
stevedore==3.5.0
threadpoolctl==3.1.0
traitlets==5.3.0
urllib3==1.26.9
userpath==1.8.0
virtualenv==20.13.4
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
websocket-client==1.3.1
winshell==0.6
